(Re)move utils (#67)
* reorganize utils

* add missing init

* fix import in test
