+++ qiskit/optimization/__init__.py
@@ -28,7 +28,7 @@ Submodules
 """
 
 from .infinity import infinity  # must be at the top of the file
-from .utils import QiskitOptimizationError
+from .exceptions import QiskitOptimizationError
 from .problems import QuadraticProgram
 from ._logging import (get_qiskit_optimization_logging,
                        set_qiskit_optimization_logging)
