+++ qiskit/optimization/algorithms/minimum_eigen_optimizer.py
@@ -15,14 +15,15 @@
 
 """A wrapper for minimum eigen solvers from Qiskit Aqua to be used within Qiskit Optimization."""
 
-from typing import Optional, Any
+from typing import Optional, Any, Union, Tuple, List
 import numpy as np
 
+from qiskit import QuantumCircuit, BasicAer, execute
 from qiskit.aqua.algorithms import MinimumEigensolver
+from qiskit.aqua.operators import WeightedPauliOperator, MatrixOperator
 
 from .optimization_algorithm import OptimizationAlgorithm
 from ..problems.quadratic_program import QuadraticProgram
-from ..utils.eigenvector_to_solutions import eigenvector_to_solutions
 from ..converters.quadratic_program_to_operator import QuadraticProgramToOperator
 from ..converters.quadratic_program_to_qubo import QuadraticProgramToQubo
 from ..results.optimization_result import OptimizationResult
@@ -151,3 +152,91 @@ class MinimumEigenOptimizer(OptimizationAlgorithm):
 
         # translate results back to original problem
         return opt_res
+
+
+def eigenvector_to_solutions(eigenvector: Union[dict, np.ndarray],
+                             operator: Union[WeightedPauliOperator, MatrixOperator],
+                             min_probability: float = 1e-6) -> List[Tuple[str, float, float]]:
+    """Convert the eigenvector to the bitstrings and corresponding eigenvalues.
+
+    Examples:
+    >>> op = MatrixOperator(numpy.array([[1, 1], [1, -1]]) / numpy.sqrt(2))
+    >>> eigenvectors = {'0': 12, '1': 1}
+    >>> print(eigenvector_to_solutions(eigenvectors, op))
+    [('0', 0.7071067811865475, 0.9230769230769231), ('1', -0.7071067811865475, 0.07692307692307693)]
+
+    >>> op = MatrixOperator(numpy.array([[1, 1], [1, -1]]) / numpy.sqrt(2))
+    >>> eigenvectors = numpy.array([1, 1] / numpy.sqrt(2), dtype=complex)
+    >>> print(eigenvector_to_solutions(eigenvectors, op))
+    [('0', 0.7071067811865475, 0.4999999999999999), ('1', -0.7071067811865475, 0.4999999999999999)]
+
+    Returns:
+        For each computational basis state contained in the eigenvector, return the basis
+        state as bitstring along with the operator evaluated at that bitstring and the
+        probability of sampling this bitstring from the eigenvector.
+
+    Raises:
+        TypeError: Invalid Argument
+
+    """
+    solutions = []
+    if isinstance(eigenvector, dict):
+        all_counts = sum(eigenvector.values())
+        # iterate over all samples
+        for bitstr, count in eigenvector.items():
+            sampling_probability = count / all_counts
+            # add the bitstring, if the sampling probability exceeds the threshold
+            if sampling_probability > 0:
+                if sampling_probability >= min_probability:
+                    value = eval_operator_at_bitstring(operator, bitstr)
+                    solutions += [(bitstr, value, sampling_probability)]
+
+    elif isinstance(eigenvector, np.ndarray):
+        num_qubits = int(np.log2(eigenvector.size))
+        probabilities = np.abs(eigenvector * eigenvector.conj())
+
+        # iterate over all states and their sampling probabilities
+        for i, sampling_probability in enumerate(probabilities):
+
+            # add the i-th state if the sampling probability exceeds the threshold
+            if sampling_probability > 0:
+                if sampling_probability >= min_probability:
+                    bitstr = '{:b}'.format(i).rjust(num_qubits, '0')[::-1]
+                    value = eval_operator_at_bitstring(operator, bitstr)
+                    solutions += [(bitstr, value, sampling_probability)]
+
+    else:
+        raise TypeError('Unsupported format of eigenvector. Provide a dict or numpy.ndarray.')
+
+    return solutions
+
+
+def eval_operator_at_bitstring(operator: Union[WeightedPauliOperator, MatrixOperator],
+                               bitstr: str) -> float:
+    """Evaluate an Aqua operator at a given bitstring.
+
+    This simulates a circuit representing the bitstring. Note that this will not be needed
+    with the Operator logic introduced in 0.7.0.
+
+    Args:
+        operator: The operator which is evaluated.
+        bitstr: The bitstring at which the operator is evaluated.
+
+    Returns:
+        The operator evaluated with the quantum state the bitstring describes.
+    """
+
+    # TODO check that operator size and bitstr are compatible
+    circuit = QuantumCircuit(len(bitstr))
+    for i, bit in enumerate(bitstr):
+        # TODO in which order to iterate over the bitstring???
+        if bit == '1':
+            circuit.x(i)
+
+    # simulate the circuit
+    result = execute(circuit, BasicAer.get_backend('statevector_simulator')).result()
+
+    # evaluate the operator
+    value = np.real(operator.evaluate_with_statevector(result.get_statevector())[0])
+
+    return value
