+++ qiskit/optimization/exceptions/qiskit_optimization_error.py
