+++ qiskit/optimization/problems/variables.py
@@ -18,9 +18,9 @@ import copy
 from typing import List, Optional, Union
 
 from qiskit.optimization import infinity
-from qiskit.optimization.utils.base import BaseInterface
-from qiskit.optimization.utils.helpers import init_list_args, NameIndex
-from qiskit.optimization.utils.qiskit_optimization_error import QiskitOptimizationError
+from qiskit.optimization.problems.name_index import init_list_args, NameIndex
+from qiskit.optimization.exceptions import QiskitOptimizationError
+from .base import BaseInterface
 
 CPX_CONTINUOUS = 'C'
 CPX_BINARY = 'B'
