+++ qiskit/optimization/results/solution.py
@@ -17,7 +17,7 @@
 from qiskit.optimization.results.quality_metrics import QualityMetrics
 from qiskit.optimization.results.solution_status import SolutionStatus
 
-from qiskit.optimization.utils.base import BaseInterface
+from ..problems.base import BaseInterface
 
 
 class SolutionInterface(BaseInterface):
